requirements.txt
image==1.5.33
PyYAML==6.0
